# docs/swagger.yaml
basePath: /
definitions:
...
  api.ErrorResponse:
    properties:
      details:
        example: '''to'' field is required'
        type: string
      error:
        example: validation failed
        type: string
    type: object
  api.MessageResponse:
    properties:
      message:
        example: operation completed successfully
        type: string
    type: object
  api.RegisterDeviceRequest:
    properties:
      platform:
        enum:
        - android
        - ios
        - web
        example: android
        type: string
      token:
        example: fcm-token-abc123
        type: string
      user_id:
        example: user-123
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@yourdomain.com
    name: API Support
  description: Unified notification delivery service supporting email, SMS, push notifications,
    and in-app messaging with bulk sending capabilities.
  title: Buzz Notification Service API
  version: 1.0.0
paths:
...
  /api/v1/devices:
    get:
      description: Retrieve all registered device tokens for a user
      parameters:
      - description: User ID
        in: query
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - Bearer: []
      summary: List user devices
      tags:
      - devices
  /api/v1/devices/{token}:
    delete:
      description: Deactivate a push notification device token
      parameters:
      - description: Device token
        in: path
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - Bearer: []
      summary: Unregister a device
      tags:
      - devices
  /api/v1/devices/register:
    post:
      consumes:
      - application/json
      description: Register a push notification device token for a user
      parameters:
      - description: Device registration payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.RegisterDeviceRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - Bearer: []
      summary: Register a device
      tags:
      - devices
  /api/v1/inbox:
    get:
      description: Retrieve the authenticated user's in-app notification inbox
      parameters:
      - description: Return only unread notifications (default false)
        in: query
        name: unread
        type: boolean
      - description: Page size (default 20)
        in: query
        name: limit
        type: integer
      - description: Page offset (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - Bearer: []
      summary: Get inbox
      tags:
      - inbox
  /api/v1/inbox/{id}:
    delete:
      description: Permanently delete a notification from the authenticated user's
        inbox
      parameters:
      - description: Inbox entry UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - Bearer: []
      summary: Delete inbox notification
      tags:
      - inbox
  /api/v1/inbox/{id}/read:
    patch:
      description: Mark a single inbox notification as read
      parameters:
      - description: Inbox entry UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - Bearer: []
      summary: Mark as read
      tags:
      - inbox
  /api/v1/inbox/read-all:
    post:
      description: Mark all inbox notifications as read for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - Bearer: []
      summary: Mark all as read
      tags:
      - inbox
  /api/v1/monitoring/queues:
    get:
      description: Retrieve the names of all active notification queues
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - Bearer: []
      summary: List queues
      tags:
      - monitoring
  /api/v1/monitoring/queues/{queue}:
    get:
      description: Retrieve processing statistics for a specific queue
      parameters:
      - description: Queue name
        in: path
        name: queue
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - Bearer: []
      summary: Get queue stats
      tags:
      - monitoring
  /api/v1/monitoring/stats:
    get:
      description: Retrieve aggregated processing statistics across all notification
        queues
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - Bearer: []
      summary: Get all queue stats
      tags:
      - monitoring
...
securityDefinitions:
  Bearer:
    description: API key prefixed with "Bearer ", e.g. "Bearer YOUR_API_KEY"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"